pyproject.toml
[project]
name = "community-persona-pipeline"
version = "0.1.0"
description = ""
authors = [
    {name = "Kyeong6",email = "wwnss1411@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp (>=1.18.0,<2.0.0)",
    "playwright (>=1.55.0,<2.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
crawl = "src.mcp_server:main"
test = "pytest"
